refactor: generalize passthrough handling
Allows for code reuse between Stemmarest and Stemweb
API endpoints
