+++ middleware/stemmaweb_middleware/resources/base/__init__.py
@@ -1,3 +1,4 @@
 from .api_client import APIClient
+from .passthrough_handler import handle_passthrough_request
 
-__all__ = ["APIClient"]
+__all__ = ["APIClient", "handle_passthrough_request"]
